requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
#backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
googlesearch-python==1.2.3
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
xmltodict==0.13.0
geopy==2.4.1
openai
psutil==5.9.7
matplotlib==3.8.2
pandas==2.1.4
django-filter-23.5